requirements.txt
Pillow
python-dotenv
discord.py
tensorflow
numpy